fix many things for react 17
