+++ client/package.json
@@ -1,23 +1,23 @@
 {
-  "name": "node-against",
+  "name": "house-of-decks-client",
   "version": "0.1.0",
   "proxy": "http://localhost:5000",
   "private": true,
   "dependencies": {
     "@emotion/core": "^10.0.35",
-    "@testing-library/jest-dom": "^5.11.4",
+    "@testing-library/jest-dom": "^5.11.5",
     "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.8",
+    "@testing-library/user-event": "^12.1.10",
     "canvas-confetti": "^1.3.1",
-    "formik": "^2.2.0",
-    "framer-motion": "^2.9.1",
+    "formik": "^2.2.1",
+    "framer-motion": "^2.9.4",
     "history": "^5.0.0",
-    "react": "^16.14.0",
-    "react-cool-portal": "^1.0.20",
-    "react-dom": "^16.14.0",
+    "react": "^17.0.1",
+    "react-cool-portal": "^1.0.23",
+    "react-dom": "^17.0.1",
     "react-responsive": "^8.1.0",
     "react-router-dom": "^6.0.0-beta.0",
-    "react-scripts": "^4.0.0-next.98",
+    "react-scripts": "^4.0.0",
     "react-use-clipboard": "^1.0.1",
     "reconnecting-websocket": "^4.4.0",
     "use-sound": "^1.0.2",
