+++ client/src/Components/DeckSelect/index.js
@@ -1,6 +1,5 @@
-/** @jsx jsx */
-import { jsx, css } from "@emotion/core";
-import { motion } from "framer-motion";
+import styled from "@emotion/styled";
+import { m as motion } from "framer-motion";
 import { SoundContext } from "Sounds/Context";
 import { useContext, useState } from "react";
 import Tick from "Components/Tick";
@@ -8,7 +7,7 @@ import Modal from "Components/Modal";
 
 import decksObject from "./decks.json";
 
-const itemStyle = css({
+const StyledItem = styled.div({
   cursor: "pointer",
   width: 360,
   padding: "20px",
@@ -19,7 +18,7 @@ const itemStyle = css({
   fontWeight: 900,
 });
 
-const miniCardsWrapper = css({
+const StyledMiniCards = styled.div({
   display: "flex",
   marginLeft: "auto",
   fontSize: 13,
@@ -27,7 +26,7 @@ const miniCardsWrapper = css({
   justifyContent: "space-between",
 });
 
-const miniCardStyle = css({
+const StyledMiniCard = styled.div({
   border: "2px solid black",
   background: "white",
   width: 14,
@@ -62,11 +61,7 @@ export default ({
   return (
     <Modal handleClose={handleClose}>
       {items.map((deck) => (
-        <div
-          css={itemStyle}
-          key={deck.key}
-          onClick={() => toggleDeck(deck.key)}
-        >
+        <StyledItem key={deck.key} onClick={() => toggleDeck(deck.key)}>
           <div style={{ paddingRight: 10, height: 28 }}>
             <Tick
               isChecked={selectedDecks[deck.key]}
@@ -74,10 +69,9 @@ export default ({
             />
           </div>
           <div style={{ width: 210 }}>{deck.title}</div>
-          <div css={miniCardsWrapper} title="black and white card counts">
+          <StyledMiniCards title="black and white card counts">
             <div style={{ display: "flex", width: 53, alignItems: "center" }}>
-              <div
-                css={miniCardStyle}
+              <StyledMiniCard
                 style={{
                   background: "black",
                 }}
@@ -85,11 +79,11 @@ export default ({
               {deck.blackCardCount}
             </div>
             <div style={{ display: "flex", width: 53, alignItems: "center" }}>
-              <div css={miniCardStyle} />
+              <StyledMiniCard />
               {deck.whiteCardCount}
             </div>
-          </div>
-        </div>
+          </StyledMiniCards>
+        </StyledItem>
       ))}
     </Modal>
   );
