+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "house-of-decks",
+  "name": "house-of-decks-server",
   "version": "1.0.0",
-  "description": "api",
+  "description": "websocket server for house of decks game",
   "main": "server/index.js",
   "engines": {
     "node": "12.16.x"
@@ -24,17 +24,19 @@
   },
   "homepage": "https://github.com/keremciu/house-of-decks#readme",
   "dependencies": {
+    "@emotion/styled": "^10.0.27",
     "esm": "^3.2.25",
     "express": "^4.17.1",
+    "react-scripts": "^4.0.0",
     "ws": "^7.3.1"
   },
   "devDependencies": {
-    "eslint": "^7.11.0",
-    "eslint-config-standard": "^14.1.1",
+    "eslint": "^7.12.0",
+    "eslint-config-standard": "^15.0.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.1",
-    "nodemon": "^2.0.5"
+    "eslint-plugin-standard": "^4.0.2",
+    "nodemon": "^2.0.6"
   }
 }
